Lab7: Base 1, 2, 3, 4, Adv 1, 2

+++ Lab6/src/kernel/reboot.c
@@ -7,13 +7,13 @@ void set(long addr, unsigned int value)
 }
 
 void reset(int tick)
-{                                     // reboot after watchdog timer expire
-    set(PM_RSTC, PM_PASSWORD | 0x20); // full reset
-    set(PM_WDOG, PM_PASSWORD | tick); // number of watchdog tick
+{                                                   // reboot after watchdog timer expire
+    set(PM_RSTC, (unsigned int)PM_PASSWORD | 0x20); // full reset
+    set(PM_WDOG, (unsigned int)PM_PASSWORD | tick); // number of watchdog tick
 }
 
 void cancel_reset()
 {
-    set(PM_RSTC, PM_PASSWORD | 0); // full reset
-    set(PM_WDOG, PM_PASSWORD | 0); // number of watchdog tick
+    set(PM_RSTC, (unsigned int)PM_PASSWORD | 0); // full reset
+    set(PM_WDOG, (unsigned int)PM_PASSWORD | 0); // number of watchdog tick
 }
